+++ src/App/appStore.ts
@@ -1,9 +1,21 @@
-import { configureStore } from '@reduxjs/toolkit'
+import { combineReducers, configureStore } from '@reduxjs/toolkit'
 import { setupListeners } from '@reduxjs/toolkit/dist/query/react'
 
-import { baseApi } from '@/Shared/Api'
 import { invalidateAccessTokenListener } from '@/Entities/Session'
-import { rootReducer } from '@/Entities/rootReducer.ts'
+
+import { baseApi } from '@/Shared/Api'
+import { sessionSlice } from '@/Entities/Session'
+import { userSlice } from '@/Entities/User'
+import { boardsSlice } from '@/Entities/Boards'
+import { tasksSlice } from '@/Entities/Tasks'
+
+export const rootReducer = combineReducers({
+    [userSlice.name]: userSlice.reducer,
+    [sessionSlice.name]: sessionSlice.reducer,
+    [boardsSlice.name]: boardsSlice.reducer,
+    [tasksSlice.name]: tasksSlice.reducer,
+    [baseApi.reducerPath]: baseApi.reducer,
+})
 
 export const store = configureStore({
     reducer: rootReducer,
